+++ src/main/java/ro/ubbcluj/apm/am/processor/NumberProcessor.java
@@ -1,0 +1,5 @@
+package ro.ubbcluj.apm.am.processor;
+
+@FunctionalInterface
+public interface NumberProcessor<T> extends Processor<T, Number> {
+}