+++ src/main/java/ro/ubbcluj/apm/am/processor/Processor.java
@@ -1,0 +1,6 @@
+package ro.ubbcluj.apm.am.processor;
+
+@FunctionalInterface
+public interface Processor<T, E> {
+    T process(E element);
+}