+++ src/main/java/ro/ubbcluj/apm/am/processor/StringProcessor.java
@@ -1,0 +1,5 @@
+package ro.ubbcluj.apm.am.processor;
+
+@FunctionalInterface
+public interface StringProcessor<T> extends Processor<T, String> {
+}