+++ src/main/java/ro/ubbcluj/apm/am/processor/Student.java
@@ -1,0 +1,10 @@
+package ro.ubbcluj.apm.am.processor;
+
+import lombok.Data;
+
+@Data
+public class Student {
+    private String name;
+    private int age;
+    private double grade;
+}