+++ src/test/java/ro/ubbcluj/apm/am/processor/StudentTest.java
@@ -1,0 +1,13 @@
+package ro.ubbcluj.apm.am.processor;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+public class StudentTest {
+
+    @Test
+    void filterStudents(){
+
+    }
+}